program_crypto/requirements.txt
pandas
numpy
python-binance
ta
requests
schedule
pyyaml
python-telegram-bot
